consolidating common robust alignment tools
