+++ engeom/src/geom2/align2/points_to_surface.rs
@@ -1,7 +1,7 @@
 use crate::Result;
 use crate::common::SPCoords;
 use crate::common::align::{RefinementHalt, SolveQuality, TerminationReason};
-use crate::common::consensus::weights::MagsacWeight;
+use crate::common::consensus::weights::{MagsacWeight, estimate_sigma_max};
 use crate::common::dist;
 use crate::geom2::Point2;
 use crate::geom2::align2::jacobian::{copy_jacobian, point_surf_jacobian2};
@@ -17,10 +17,6 @@ use levenberg_marquardt::{LeastSquaresProblem, LevenbergMarquardt};
 /// degrees of freedom. (`MagsacWeight` requires at least 2; a point-to-plane residual would be a
 /// one-dimensional projection and would need the weight function extended.)
 const RESIDUAL_DOF: usize = 2;
-
-/// The scale factor that turns a median absolute deviation into a consistent estimate of the
-/// standard deviation of normally distributed data.
-const MAD_TO_SIGMA: f64 = 1.4826;
 
 /// Performs a Levenberg-Marquardt minimization to align a set of 2D points to a surface target.
 ///
@@ -201,43 +197,6 @@ fn resolve_sigma_max<T: SurfaceTarget2>(
         Some(s) => Some(s),
         None => estimate_sigma_max(&problem.normalized_residuals()),
     }
-}
-
-/// Estimates a MAGSAC++ `sigma_max` from a set of residuals via the median absolute deviation.
-///
-/// MAD is used rather than the standard deviation because it is insensitive to the gross outliers
-/// the robust weighting exists to suppress: contaminating up to half the data cannot move it
-/// arbitrarily, whereas a single distant point can dominate a standard deviation.
-///
-/// Returns `None` when the spread is zero or non-finite, which happens when the fit is already
-/// essentially exact and there is nothing to reweight.
-fn estimate_sigma_max(residuals: &[f64]) -> Option<f64> {
-    let center = median(residuals)?;
-    let deviations: Vec<f64> = residuals.iter().map(|r| (r - center).abs()).collect();
-    let sigma = MAD_TO_SIGMA * median(&deviations)?;
-
-    if sigma.is_finite() && sigma > 0.0 {
-        Some(sigma)
-    } else {
-        None
-    }
-}
-
-/// The median of a slice of finite values, or `None` if the slice is empty.
-fn median(values: &[f64]) -> Option<f64> {
-    if values.is_empty() {
-        return None;
-    }
-
-    let mut sorted = values.to_vec();
-    sorted.sort_by(|a, b| a.total_cmp(b));
-
-    let n = sorted.len();
-    Some(if n.is_multiple_of(2) {
-        0.5 * (sorted[n / 2 - 1] + sorted[n / 2])
-    } else {
-        sorted[n / 2]
-    })
 }
 
 struct PointsToSurface2<'a, T: SurfaceTarget2> {
@@ -1159,34 +1118,4 @@ mod tests {
     // ============================================================================================
     // Supporting pieces
     // ============================================================================================
-
-    #[test]
-    fn median_handles_both_parities() {
-        assert_eq!(median(&[3.0, 1.0, 2.0]), Some(2.0));
-        assert_eq!(median(&[4.0, 1.0, 3.0, 2.0]), Some(2.5));
-        assert_eq!(median(&[]), None);
-    }
-
-    #[test]
-    fn sigma_estimate_is_robust_to_outliers() {
-        // Eleven values with unit spread, plus one enormous outlier. A standard deviation would
-        // be dominated by the outlier; the MAD-based estimate should barely notice it.
-        let mut values: Vec<f64> = (-5..=5).map(|i| i as f64).collect();
-        let clean = estimate_sigma_max(&values).unwrap();
-
-        values.push(10_000.0);
-        let contaminated = estimate_sigma_max(&values).unwrap();
-
-        assert!(
-            (contaminated - clean).abs() < 0.5 * clean,
-            "outlier moved the estimate too far: {clean} -> {contaminated}"
-        );
-    }
-
-    #[test]
-    fn sigma_estimate_rejects_degenerate_spread() {
-        // Every residual identical means no spread to estimate from.
-        assert_eq!(estimate_sigma_max(&[2.0; 10]), None);
-        assert_eq!(estimate_sigma_max(&[]), None);
-    }
 }
